requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas==2.1.3
numpy==1.26.2
openai>=1.12.0
httpx>=0.25.0
python-dotenv==1.0.0
gunicorn==21.2.0
